fix: harden M2 final cutover evidence
